Run integration tests via cargo run
