+++ tests/Cargo.toml
@@ -9,11 +9,7 @@ homepage = "https://solana.com/"
 edition = "2018"
 
 [dependencies]
-
-[dev-dependencies]
 hex = "0.4.1"
-serial_test = "0.3.2"
-serial_test_derive = "0.3.1"
 solana-cli = "1.0.0"
 solana-remote-wallet = "1.0.0"
 solana-sdk = "1.0.0"
